+++ src/utils/suggester-workers/commons-tokenizer/prepare-string-indexation.ts
